--- components/Notes/test.json
@@ -1,6 +1,0 @@
-[
-  { "name": "213", "added": false, "id": "5NQlmO" },
-  { "name": "4342432", "added": false, "id": "L5K72t" },
-  { "name": "314", "added": false, "id": "wpYdUA" },
-  { "name": "2", "added": false, "id": "hf_f4B" }
-]
