+++ jest.config.js
@@ -13,7 +13,7 @@ const customJestConfig = {
   // if using TypeScript with a baseUrl set to the root directory then you need the below for alias' to work
   moduleDirectories: ["node_modules", "<rootDir>/"],
   testEnvironment: "jest-environment-jsdom",
-  testMatch: ["<rootDir>/components/**/*.test.js"],
+  testMatch: ["<rootDir>/Components/**/*.test.js"],
   // add path alias (~ for root directory)
   moduleNameMapper: {
     "^~/(.*)$": "<rootDir>/$1",
